requirements.txt
pandas
requests
supabase
gspread
gspread-dataframe
python-dotenv
openpyxl